Extract environment setup from file system tests
Extract the environment setup from tests that override
AbstractTestTrinoFileSystem so that the environment can
be reused for other tests.
