+++ lib/trino-filesystem-s3/src/test/java/io/trino/filesystem/s3/TestS3FileSystemLocalStack.java
@@ -13,60 +13,33 @@
  */
 package io.trino.filesystem.s3;
 
-import io.airlift.units.DataSize;
-import org.testcontainers.containers.localstack.LocalStackContainer;
-import org.testcontainers.containers.localstack.LocalStackContainer.Service;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
+import io.trino.filesystem.s3.S3FileSystemTestingEnvironment.S3FileSystemTestingEnvironmentLocalStack;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 
-@Testcontainers
 public class TestS3FileSystemLocalStack
         extends AbstractTestS3FileSystem
 {
-    private static final String BUCKET = "test-bucket";
+    private S3FileSystemTestingEnvironmentLocalStack testingEnvironment;
 
-    @Container
-    private static final LocalStackContainer LOCALSTACK = new LocalStackContainer(DockerImageName.parse("localstack/localstack:2.0.2"))
-            .withServices(Service.S3);
-
-    @Override
-    protected void initEnvironment()
+    @BeforeAll
+    public void setup()
     {
-        try (S3Client s3Client = createS3Client()) {
-            s3Client.createBucket(builder -> builder.bucket(BUCKET).build());
+        testingEnvironment = new S3FileSystemTestingEnvironmentLocalStack();
+    }
+
+    @AfterAll
+    public void cleanup()
+    {
+        if (testingEnvironment != null) {
+            testingEnvironment.close();
+            testingEnvironment = null;
         }
     }
 
     @Override
-    protected String bucket()
+    protected S3FileSystemTestingEnvironment s3TestingEnvironment()
     {
-        return BUCKET;
-    }
-
-    @Override
-    protected S3Client createS3Client()
-    {
-        return S3Client.builder()
-                .endpointOverride(LOCALSTACK.getEndpointOverride(Service.S3))
-                .region(Region.of(LOCALSTACK.getRegion()))
-                .credentialsProvider(StaticCredentialsProvider.create(
-                        AwsBasicCredentials.create(LOCALSTACK.getAccessKey(), LOCALSTACK.getSecretKey())))
-                .build();
-    }
-
-    @Override
-    protected S3FileSystemFactory createS3FileSystemFactory()
-    {
-        return new S3FileSystemFactory(new S3FileSystemConfig()
-                .setAwsAccessKey(LOCALSTACK.getAccessKey())
-                .setAwsSecretKey(LOCALSTACK.getSecretKey())
-                .setEndpoint(LOCALSTACK.getEndpointOverride(Service.S3).toString())
-                .setRegion(LOCALSTACK.getRegion())
-                .setStreamingPartSize(DataSize.valueOf("5.5MB")));
+        return testingEnvironment;
     }
 }
