+++ lib/trino-filesystem/src/test/java/io/trino/filesystem/local/TestLocalFileSystem.java
@@ -14,97 +14,49 @@
 package io.trino.filesystem.local;
 
 import io.trino.filesystem.AbstractTestTrinoFileSystem;
-import io.trino.filesystem.Location;
-import io.trino.filesystem.TrinoFileSystem;
+import io.trino.filesystem.AbstractTrinoFileSystemTestingEnvironment;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Comparator;
-import java.util.Iterator;
-import java.util.stream.Stream;
 
-import static java.util.function.Predicate.not;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class TestLocalFileSystem
         extends AbstractTestTrinoFileSystem
 {
-    private LocalFileSystem fileSystem;
-    private Path tempDirectory;
+    private LocalFileSystemTestingEnvironment testingEnvironment;
 
     @BeforeAll
     void beforeAll()
             throws IOException
     {
-        tempDirectory = Files.createTempDirectory("test");
-        fileSystem = new LocalFileSystem(tempDirectory);
+        testingEnvironment = new LocalFileSystemTestingEnvironment();
     }
 
     @AfterEach
     void afterEach()
             throws IOException
     {
-        cleanupFiles();
+        testingEnvironment.cleanupFiles();
     }
 
     @AfterAll
     void afterAll()
             throws IOException
     {
-        Files.delete(tempDirectory);
-    }
-
-    private void cleanupFiles()
-            throws IOException
-    {
-        // tests will leave directories
-        try (Stream<Path> walk = Files.walk(tempDirectory)) {
-            Iterator<Path> iterator = walk.sorted(Comparator.reverseOrder()).iterator();
-            while (iterator.hasNext()) {
-                Path path = iterator.next();
-                if (!path.equals(tempDirectory)) {
-                    Files.delete(path);
-                }
-            }
+        if (testingEnvironment != null) {
+            testingEnvironment.close();
+            testingEnvironment = null;
         }
     }
 
     @Override
-    protected boolean isHierarchical()
+    protected AbstractTrinoFileSystemTestingEnvironment testingEnvironment()
     {
-        return true;
-    }
-
-    @Override
-    protected TrinoFileSystem getFileSystem()
-    {
-        return fileSystem;
-    }
-
-    @Override
-    protected Location getRootLocation()
-    {
-        return Location.of("local://");
-    }
-
-    @Override
-    protected void verifyFileSystemIsEmpty()
-    {
-        try {
-            try (Stream<Path> entries = Files.list(tempDirectory)) {
-                assertThat(entries.filter(not(tempDirectory::equals)).findFirst()).isEmpty();
-            }
-        }
-        catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return testingEnvironment;
     }
 
     @Test
